requirements.txt
Flask==3.1.1
Flask-Admin==1.6.1
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
python-dotenv==1.1.0
SQLAlchemy==2.0.41
WTForms==3.2.1
alembic==1.16.1
bcrypt==4.3.0
blinker==1.9.0
click==8.3.0
greenlet==3.2.2
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
PyJWT==2.10.1
typing_extensions==4.13.2
Werkzeug==3.1.3
